Use a src directory structure for library
Split the tooling across the init file and an CLI file
